Добавляет модуляцию js-файлов
