+++ index.html
@@ -241,8 +241,8 @@
       </div>
     </section>
   </template>
-  <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
+  <!-- <script src="/js/functions.js"></script> -->
+  <script src="/js/main.js" type="module"></script>
 </body>
 
 </html>
