+++ js/constants.js
@@ -1,0 +1,60 @@
+//сколько объектов нужно сгенерировать в массиве photos
+const PHOTOS = 25;
+
+//минимальное и максимальное значение лайков для генерации числа лайков под фото
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+
+//общее количество комментариев
+const COMMENTS = 100;
+//минимальное количестко комментариев под фото
+const MIN_COMMENTS = 0;
+//максимальное количестко комментариев под фото
+const MAX_COMMENTS = 30;
+
+//для генерации аватара коментирующего
+const MIN_AVATAR = 1;
+const MAX_AVATAR = 6;
+
+//массив, с описаниями к фото
+const DESCRIPTIONS = [
+  'На море',
+  'Котик',
+  'Путешествуем',
+  'Моя ласточка',
+  'На концерте',
+  'Пальчики оближешь',
+];
+
+//массив, с комментариями к фото
+const MESSAGES = [
+  'Красиво',
+  'Наверно вкусно',
+  'А это где?',
+  'Вот это аппарат',
+  'Что за группа?',
+  'Это можно есть?',
+];
+
+//массив, с именами коментирующих фото
+const NAMES = [
+  'Оля',
+  'Никита',
+  'Света',
+  'Роман',
+  'Тоня',
+  'Олег',
+];
+export {
+  PHOTOS,
+  MIN_LIKES,
+  MAX_LIKES,
+  COMMENTS,
+  MIN_COMMENTS,
+  MAX_COMMENTS,
+  MIN_AVATAR,
+  MAX_AVATAR,
+  DESCRIPTIONS,
+  MESSAGES,
+  NAMES
+};
