+++ js/utils.js
@@ -1,0 +1,30 @@
+//функции общего пользования
+//генерация случайного числа без контроля повторений
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// генерация случайного числа с контролем повторений,
+// функция с использованием замыкания( функция возвращает функцию, которая написана с лексическим окружением)
+const createRandomIdFromRangeGenerator = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+export {
+  getRandomInteger,
+  createRandomIdFromRangeGenerator
+};
